[PC-425] refactor: AuthenticationService 모듈 위치 수정 common -> common:auth
