--- common/src/main/java/org/yapp/application/AuthenticationService.java
@@ -1,16 +1,0 @@
-package org.yapp.application;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class AuthenticationService {
-  public Authentication getAuthentication() {
-    return SecurityContextHolder.getContext().getAuthentication();
-  }
-
-  public Long getUserId() {
-    return (Long) getAuthentication().getPrincipal();
-  }
-}
